Moves new methods to template method class
