+++ artemis-core/src/main/java/org/jnosql/artemis/column/query/ColumnRepositoryProxy.java
@@ -18,27 +18,11 @@ package org.jnosql.artemis.column.query;
 
 import org.jnosql.artemis.Repository;
 import org.jnosql.artemis.column.ColumnTemplate;
-import org.jnosql.artemis.key.KeyNotFoundException;
 import org.jnosql.artemis.reflection.ClassRepresentation;
 import org.jnosql.artemis.reflection.ClassRepresentations;
-import org.jnosql.artemis.reflection.FieldRepresentation;
 import org.jnosql.artemis.reflection.Reflections;
-import org.jnosql.diana.api.column.Column;
-import org.jnosql.diana.api.column.ColumnCondition;
-import org.jnosql.diana.api.column.ColumnDeleteQuery;
-import org.jnosql.diana.api.column.ColumnQuery;
 
 import java.lang.reflect.ParameterizedType;
-import java.util.Objects;
-import java.util.Optional;
-import java.util.function.Function;
-import java.util.function.Supplier;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-import java.util.stream.StreamSupport;
-
-import static java.util.Objects.requireNonNull;
-import static java.util.stream.StreamSupport.stream;
 
 
 /**
@@ -48,8 +32,6 @@ import static java.util.stream.StreamSupport.stream;
  */
 class ColumnRepositoryProxy<T, ID> extends AbstractColumnRepositoryProxy {
 
-    private static final Supplier<KeyNotFoundException> KEY_NOT_FOUND_EXCEPTION_SUPPLIER = ()
-            -> new KeyNotFoundException("To use this resource you must annotaded a fiel with @org.jnosql.artemisId");
 
     private final Class<T> typeClass;
 
@@ -109,12 +91,9 @@ class ColumnRepositoryProxy<T, ID> extends AbstractColumnRepositoryProxy {
 
         private final ClassRepresentation classRepresentation;
 
-        private final Optional<FieldRepresentation> idField;
-
         ColumnRepository(ColumnTemplate template, ClassRepresentation classRepresentation) {
             this.template = template;
             this.classRepresentation = classRepresentation;
-            this.idField = classRepresentation.getId();
         }
 
         @Override
@@ -123,61 +102,14 @@ class ColumnRepositoryProxy<T, ID> extends AbstractColumnRepositoryProxy {
         }
 
         @Override
-        public void deleteById(Object id) throws NullPointerException {
-            requireNonNull(id, "is is required");
-            ColumnDeleteQuery query = ColumnDeleteQuery.of(classRepresentation.getName());
-            String columnName = this.idField.orElseThrow(KEY_NOT_FOUND_EXCEPTION_SUPPLIER).getName();
-            query.with(ColumnCondition.eq(Column.of(columnName, id)));
+        protected ClassRepresentation getClassRepresentation() {
+            return classRepresentation;
         }
 
         @Override
-        public void deleteById(Iterable ids) throws NullPointerException {
-            requireNonNull(ids, "ids is required");
-            ids.forEach(this::deleteById);
+        protected Reflections getReflections() {
+            return reflections;
         }
 
-        @Override
-        public void delete(Iterable entities) throws NullPointerException {
-            requireNonNull(entities, "entities is required");
-            entities.forEach(this::delete);
-        }
-
-        @Override
-        public void delete(Object entity) throws NullPointerException {
-            requireNonNull(entity, "entity is required");
-            Object idValue = reflections.getValue(entity, idField.orElseThrow(KEY_NOT_FOUND_EXCEPTION_SUPPLIER)
-                    .getField());
-            requireNonNull(idValue, "id value is required");
-            deleteById(idValue);
-        }
-
-        @Override
-        public Optional findById(Object id) throws NullPointerException {
-            requireNonNull(id, "id is required");
-
-            ColumnQuery query = ColumnQuery.of(classRepresentation.getName());
-            String columnName = this.idField.orElseThrow(KEY_NOT_FOUND_EXCEPTION_SUPPLIER).getName();
-            query.with(ColumnCondition.eq(Column.of(columnName, id)));
-            return template.singleResult(query);
-        }
-
-        @Override
-        public Iterable findById(Iterable ids) throws NullPointerException {
-            requireNonNull(ids, "ids is required");
-            return (Iterable) stream(ids.spliterator(), false)
-                    .flatMap(optionalToStream()).collect(Collectors.toList());
-        }
-
-        private Function optionalToStream() {
-            return id -> {
-                Optional entity = this.findById(id);
-                return entity.isPresent() ? Stream.of(entity.get()) : Stream.empty();
-            };
-        }
-
-        @Override
-        public boolean existsById(Object id) throws NullPointerException {
-            return findById(id).isPresent();
-        }
     }
 }
